octeontx2-pf: Add devlink support to configure TL1 RR_PRIO
In the current implementation, The TL1 RR_PRIO value is static
and kernel PF/VFs use the same value to configure its TL2 node
priority in case of DWRR children. This patch adds support to
configure TL1 RR_PRIO value using devlink. The TL1 RR_PRIO can be
configured for each PF. The VFs are not allowed to configure TL1
RR_PRIO value. The VFs can get the RR_PRIO value from the mailbox
NIX_TXSCH_ALLOC response parameter aggr_lvl_rr_prio.

Example command to configure TL1 RR_PRIO to 6 for PF 0002:04:00.0:
$ devlink -p dev param set pci/0002:04:00.0 name tl1_rr_prio value 6 \
  cmode runtime

Limitations:
1. The RR_PRIO can only be configured before VFs are enabled


Orabug: 39318904

+++ drivers/net/ethernet/marvell/octeontx2/nic/otx2_common.c
@@ -628,7 +628,7 @@ int otx2_txschq_config(struct otx2_nic *pfvf, int lvl, int prio, bool txschq_for
 	u16 (*schq_list)[MAX_TXSCHQ_PER_FUNC];
 	struct otx2_hw *hw = &pfvf->hw;
 	struct nix_txschq_config *req;
-	u64 schq, parent;
+	u16 schq, parent;
 	u64 dwrr_val;
 
 	dwrr_val = mtu_to_dwrr_weight(pfvf, pfvf->tx_max_pktlen);
